+++ exonum/src/api/backends/actix.rs
@@ -13,6 +13,9 @@
 // limitations under the License.
 
 //! Actix-web API backend.
+//!
+//! [Actix-web](https://github.com/actix/actix-web) is an asynchronous backend
+//! for HTTP API, based on the [Actix](https://github.com/actix/actix) framework.
 
 pub use actix_web::middleware::cors::Cors;
 
@@ -57,7 +60,7 @@ type SystemAddr = Addr<Syn, System>;
 pub struct RequestHandler {
     /// Endpoint name.
     pub name: String,
-    /// Endpoint http method.
+    /// Endpoint HTTP method.
     pub method: actix_web::http::Method,
     /// Inner handler.
     pub inner: Arc<RawHandler>,
@@ -263,14 +266,14 @@ pub(crate) fn create_app(aggregator: &ApiAggregator, runtime_config: ApiRuntimeC
 pub struct ApiRuntimeConfig {
     /// The socket address to bind.
     pub listen_address: SocketAddr,
-    /// Api access level.
+    /// API access level.
     pub access: ApiAccess,
     /// Optional App configuration.
     pub app_config: Option<AppConfig>,
 }
 
 impl ApiRuntimeConfig {
-    /// Creates api runtime configuration for the given address and access level.
+    /// Creates API runtime configuration for the given address and access level.
     pub fn new(listen_address: SocketAddr, access: ApiAccess) -> Self {
         Self {
             listen_address,
@@ -315,7 +318,7 @@ impl SystemRuntimeConfig {
 
 impl SystemRuntime {
     fn new(config: SystemRuntimeConfig) -> result::Result<Self, failure::Error> {
-        // Creates system thread.
+        // Creates a system thread.
         let (system_address_tx, system_address_rx) = mpsc::channel();
         let (api_runtime_tx, api_runtime_rx) = mpsc::channel();
         let api_runtimes = config.api_runtimes.clone();
@@ -412,9 +415,9 @@ impl fmt::Debug for SystemRuntime {
 /// CORS header specification.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum AllowOrigin {
-    /// Allow access from any host.
+    /// Allows access from any host.
     Any,
-    /// Allow access only from the following hosts.
+    /// Allows access only from the specified hosts.
     Whitelist(Vec<String>),
 }