+++ exonum/src/events/tests.rs
@@ -21,16 +21,15 @@ use std::{
 };
 
 use blockchain::ConsensusConfig;
-use crypto::{gen_keypair, gen_keypair_from_seed, PublicKey, Seed, Signature, SEED_LENGTH};
+use crypto::{gen_keypair, PublicKey, SecretKey};
+use env_logger;
 use events::{
     error::log_error, network::{NetworkConfiguration, NetworkPart}, noise::HandshakeParams,
     NetworkEvent, NetworkRequest,
 };
 use helpers::user_agent;
 use messages::{Connect, Message, MessageWriter, RawMessage};
-use node::{EventsPoolCapacity, NodeChannel};
-
-static FAKE_SEED: [u8; SEED_LENGTH] = [1; SEED_LENGTH];
+use node::{state::SharedConnectList, ConnectInfo, ConnectList, EventsPoolCapacity, NodeChannel};
 
 #[derive(Debug)]
 pub struct TestHandler {
@@ -61,20 +60,6 @@ impl TestHandler {
         Ok(event)
     }
 
-    pub fn connect_with(&self, addr: SocketAddr) {
-        let connect = connect_message(self.listen_address);
-        let (public_key, _) = gen_keypair_from_seed(&Seed::new(FAKE_SEED));
-        self.network_requests_tx
-            .clone()
-            .send(NetworkRequest::SendMessage(
-                addr,
-                connect.raw().clone(),
-                public_key,
-            ))
-            .wait()
-            .unwrap();
-    }
-
     pub fn disconnect_with(&self, addr: SocketAddr) {
         self.network_requests_tx
             .clone()
@@ -83,11 +68,18 @@ impl TestHandler {
             .unwrap();
     }
 
-    pub fn send_to(&self, addr: SocketAddr, raw: RawMessage) {
-        let (public_key, _) = gen_keypair_from_seed(&Seed::new(FAKE_SEED));
+    pub fn connect_with(&self, addr: SocketAddr, connect: Connect) {
         self.network_requests_tx
             .clone()
-            .send(NetworkRequest::SendMessage(addr, raw, public_key))
+            .send(NetworkRequest::SendMessage(addr, connect.raw().clone()))
+            .wait()
+            .unwrap();
+    }
+
+    pub fn send_to(&self, addr: SocketAddr, raw: RawMessage) {
+        self.network_requests_tx
+            .clone()
+            .send(NetworkRequest::SendMessage(addr, raw))
             .wait()
             .unwrap();
     }
@@ -150,13 +142,11 @@ impl TestEvents {
         }
     }
 
-    pub fn spawn(self) -> TestHandler {
-        let (mut handler_part, network_part) = self.into_reactor();
+    pub fn spawn(self, handshake_params: &HandshakeParams, connect: Connect) -> TestHandler {
+        let (mut handler_part, network_part) = self.into_reactor(connect);
+        let handshake_params = handshake_params.clone();
         let handle = thread::spawn(move || {
             let mut core = Core::new().unwrap();
-            let (public_key, secret_key) = gen_keypair_from_seed(&Seed::new(FAKE_SEED));
-            let handshake_params =
-                HandshakeParams::new(public_key, secret_key, network_part.max_message_len);
             let fut = network_part.run(&core.handle(), &handshake_params);
             core.run(fut).map_err(log_error).unwrap();
         });
@@ -164,14 +154,14 @@ impl TestEvents {
         handler_part
     }
 
-    fn into_reactor(self) -> (TestHandler, NetworkPart) {
+    fn into_reactor(self, connect: Connect) -> (TestHandler, NetworkPart) {
         let channel = NodeChannel::new(&self.events_config);
         let network_config = self.network_config;
         let (network_tx, network_rx) = channel.network_events;
         let network_requests_tx = channel.network_requests.0.clone();
 
         let network_part = NetworkPart {
-            our_connect_message: connect_message(self.listen_address),
+            our_connect_message: connect,
             listen_address: self.listen_address,
             network_config,
             max_message_len: ConsensusConfig::DEFAULT_MAX_MESSAGE_LEN,
@@ -184,14 +174,18 @@ impl TestEvents {
     }
 }
 
-pub fn connect_message(addr: SocketAddr) -> Connect {
+pub fn connect_message(
+    addr: SocketAddr,
+    public_key: &PublicKey,
+    secret_key: &SecretKey,
+) -> Connect {
     let time = time::UNIX_EPOCH;
-    Connect::new_with_signature(
-        &PublicKey::zero(),
+    Connect::new(
+        public_key,
         addr,
         time.into(),
         &user_agent::get(),
-        &Signature::zero(),
+        secret_key,
     )
 }
 
@@ -200,25 +194,73 @@ pub fn raw_message(id: u16, len: usize) -> RawMessage {
     RawMessage::new(writer.sign(&gen_keypair().1))
 }
 
+#[derive(Debug, Clone)]
+pub struct ConnectionParams {
+    pub connect: Connect,
+    pub connect_info: ConnectInfo,
+    address: SocketAddr,
+    public_key: PublicKey,
+    secret_key: SecretKey,
+    handshake_params: HandshakeParams,
+}
+
+impl ConnectionParams {
+    pub fn from_address(address: SocketAddr) -> Self {
+        let (public_key, secret_key) = gen_keypair();
+        let connect = connect_message(address, &public_key, &secret_key);
+        let handshake_params = HandshakeParams::new(
+            public_key,
+            secret_key.clone(),
+            SharedConnectList::default(),
+            ConsensusConfig::DEFAULT_MAX_MESSAGE_LEN,
+        );
+        let connect_info = ConnectInfo {
+            address,
+            public_key,
+        };
+
+        ConnectionParams {
+            connect,
+            address,
+            public_key,
+            secret_key,
+            handshake_params,
+            connect_info,
+        }
+    }
+
+    pub fn spawn(&mut self, events: TestEvents, connect_list: SharedConnectList) -> TestHandler {
+        self.handshake_params.connect_list = connect_list.clone();
+        events.spawn(&self.handshake_params, self.connect.clone())
+    }
+}
+
 #[test]
 fn test_network_handshake() {
     let first = "127.0.0.1:17230".parse().unwrap();
     let second = "127.0.0.1:17231".parse().unwrap();
 
+    let mut connect_list = ConnectList::default();
+
+    let mut t1 = ConnectionParams::from_address(first);
+    connect_list.add(t1.connect_info);
+
+    let mut t2 = ConnectionParams::from_address(second);
+    connect_list.add(t2.connect_info);
+
+    let connect_list = SharedConnectList::from_connect_list(connect_list);
+
     let e1 = TestEvents::with_addr(first);
     let e2 = TestEvents::with_addr(second);
 
-    let c1 = connect_message(first);
-    let c2 = connect_message(second);
+    let mut e1 = t1.spawn(e1, connect_list.clone());
+    let mut e2 = t2.spawn(e2, connect_list);
 
-    let mut e1 = e1.spawn();
-    let mut e2 = e2.spawn();
+    e1.connect_with(second, t1.connect.clone());
+    assert_eq!(e2.wait_for_connect(), t1.connect.clone());
 
-    e1.connect_with(second);
-    assert_eq!(e2.wait_for_connect(), c1);
-
-    e2.connect_with(first);
-    assert_eq!(e1.wait_for_connect(), c2);
+    e2.connect_with(first, t2.connect.clone());
+    assert_eq!(e1.wait_for_connect(), t2.connect.clone());
 
     e1.disconnect_with(second);
     assert_eq!(e1.wait_for_disconnect(), second);
@@ -235,16 +277,26 @@ fn test_network_big_message() {
     let m1 = raw_message(15, 100000);
     let m2 = raw_message(16, 400);
 
+    let mut connect_list = ConnectList::default();
+
+    let mut t1 = ConnectionParams::from_address(first);
+    connect_list.add(t1.connect_info);
+
+    let mut t2 = ConnectionParams::from_address(second);
+    connect_list.add(t2.connect_info);
+
+    let connect_list = SharedConnectList::from_connect_list(connect_list);
+
     let e1 = TestEvents::with_addr(first);
     let e2 = TestEvents::with_addr(second);
 
-    let mut e1 = e1.spawn();
-    let mut e2 = e2.spawn();
+    let mut e1 = t1.spawn(e1, connect_list.clone());
+    let mut e2 = t2.spawn(e2, connect_list);
 
-    e1.connect_with(second);
+    e1.connect_with(second, t1.connect.clone());
     e2.wait_for_connect();
 
-    e2.connect_with(first);
+    e2.connect_with(first, t2.connect.clone());
     e1.wait_for_connect();
 
     e1.send_to(second, m1.clone());
@@ -283,16 +335,23 @@ fn test_network_max_message_len() {
     let acceptable_message = raw_message(15, max_payload_length);
     let too_big_message = raw_message(16, max_payload_length + 1000);
 
+    let mut connect_list = ConnectList::default();
+    let mut t1 = ConnectionParams::from_address(first);
+    connect_list.add(t1.connect_info);
+    let mut t2 = ConnectionParams::from_address(second);
+    connect_list.add(t2.connect_info);
+    let connect_list = SharedConnectList::from_connect_list(connect_list);
+
     let e1 = TestEvents::with_addr(first);
     let e2 = TestEvents::with_addr(second);
 
-    let mut e1 = e1.spawn();
-    let mut e2 = e2.spawn();
+    let mut e1 = t1.spawn(e1, connect_list.clone());
+    let mut e2 = t2.spawn(e2, connect_list);
 
-    e1.connect_with(second);
+    e1.connect_with(second, t1.connect.clone());
     e2.wait_for_connect();
 
-    e2.connect_with(first);
+    e2.connect_with(first, t2.connect.clone());
     e1.wait_for_connect();
 
     e1.send_to(second, acceptable_message.clone());
@@ -308,39 +367,51 @@ fn test_network_reconnect() {
     let second = "127.0.0.1:19101".parse().unwrap();
 
     let msg = raw_message(11, 1000);
-    let c1 = connect_message(first);
 
-    let mut t1 = TestEvents::with_addr(first).spawn();
+    let mut connect_list = ConnectList::default();
+    let mut t1 = ConnectionParams::from_address(first);
+    connect_list.add(t1.connect_info);
+
+    let mut t2 = ConnectionParams::from_address(second);
+    connect_list.add(t2.connect_info);
+    let connect_list = SharedConnectList::from_connect_list(connect_list);
+
+    let e1 = TestEvents::with_addr(first);
+    let e2 = TestEvents::with_addr(second);
+
+    let mut e1 = t1.spawn(e1, connect_list.clone());
 
     // First connect attempt.
-    let mut t2 = TestEvents::with_addr(second).spawn();
+    let mut e2 = t2.spawn(e2, connect_list.clone());
 
     // Handle first attempt.
-    t1.connect_with(second);
-    assert_eq!(t2.wait_for_connect(), c1);
+    e1.connect_with(second, t1.connect.clone());
+    assert_eq!(e2.wait_for_connect(), t1.connect.clone());
 
-    t1.send_to(second, msg.clone());
-    assert_eq!(t2.wait_for_message(), msg);
+    e1.send_to(second, msg.clone());
+    assert_eq!(e2.wait_for_message(), msg);
 
-    t1.disconnect_with(second);
-    drop(t2);
-    assert_eq!(t1.wait_for_disconnect(), second);
+    e1.disconnect_with(second);
+    drop(e2);
+    assert_eq!(e1.wait_for_disconnect(), second);
 
     // Handle second attempt.
-    let mut t2 = TestEvents::with_addr(second).spawn();
+    let e2 = TestEvents::with_addr(second);
+    let mut e2 = t2.spawn(e2, connect_list);
 
-    t1.connect_with(second);
-    assert_eq!(t2.wait_for_connect(), c1);
+    e1.connect_with(second, t1.connect.clone());
+    assert_eq!(e2.wait_for_connect(), t1.connect.clone());
 
-    t1.send_to(second, msg.clone());
-    assert_eq!(t2.wait_for_message(), msg);
+    e1.send_to(second, msg.clone());
+    assert_eq!(e2.wait_for_message(), msg);
 
-    t1.disconnect_with(second);
-    assert_eq!(t1.wait_for_disconnect(), second);
+    e1.disconnect_with(second);
+    assert_eq!(e1.wait_for_disconnect(), second);
 }
 
 #[test]
 fn test_network_multiple_connect() {
+    env_logger::init();
     let main = "127.0.0.1:19600".parse().unwrap();
 
     let nodes = [
@@ -349,20 +420,49 @@ fn test_network_multiple_connect() {
         "127.0.0.1:19603".parse().unwrap(),
     ];
 
-    let mut node = TestEvents::with_addr(main).spawn();
+    let mut connect_list = ConnectList::default();
 
-    let connect_messages: Vec<_> = nodes.iter().cloned().map(connect_message).collect();
-    let connectors: Vec<_> = nodes
+    let mut connection_params: Vec<_> = nodes
         .iter()
-        .map(|addr| TestEvents::with_addr(*addr).spawn())
+        .cloned()
+        .map(|addr| ConnectionParams::from_address(addr))
         .collect();
 
-    connectors[0].connect_with(main);
-    assert_eq!(node.wait_for_connect(), connect_messages[0]);
-    connectors[1].connect_with(main);
-    assert_eq!(node.wait_for_connect(), connect_messages[1]);
-    connectors[2].connect_with(main);
-    assert_eq!(node.wait_for_connect(), connect_messages[2]);
+    for params in connection_params.iter().cloned() {
+        connect_list.add(params.connect_info);
+    }
+
+    let mut t1 = ConnectionParams::from_address(main);
+    let events = TestEvents::with_addr(t1.address);
+
+    connect_list.add(t1.connect_info);
+
+    let connect_list = SharedConnectList::from_connect_list(connect_list);
+    let mut node = t1.spawn(events, connect_list.clone());
+
+    let connectors: Vec<_> = connection_params
+        .iter_mut()
+        .map(|params| {
+            let events = TestEvents::with_addr(params.address);
+            params.spawn(events, connect_list.clone())
+        })
+        .collect();
+
+    connectors[0].connect_with(main, connection_params[0].connect.clone());
+    assert_eq!(
+        node.wait_for_connect(),
+        connection_params[0].connect.clone()
+    );
+    connectors[1].connect_with(main, connection_params[1].connect.clone());
+    assert_eq!(
+        node.wait_for_connect(),
+        connection_params[1].connect.clone()
+    );
+    connectors[2].connect_with(main, connection_params[2].connect.clone());
+    assert_eq!(
+        node.wait_for_connect(),
+        connection_params[2].connect.clone()
+    );
 }
 
 #[test]
@@ -370,12 +470,22 @@ fn test_send_first_not_connect() {
     let main = "127.0.0.1:19500".parse().unwrap();
     let other = "127.0.0.1:19501".parse().unwrap();
 
-    let mut node = TestEvents::with_addr(main).spawn();
-    let other_node = TestEvents::with_addr(other).spawn();
+    let mut connect_list = ConnectList::default();
+    let mut t1 = ConnectionParams::from_address(main);
+    connect_list.add(t1.connect_info);
+    let mut t2 = ConnectionParams::from_address(other);
+    connect_list.add(t2.connect_info);
+    let connect_list = SharedConnectList::from_connect_list(connect_list);
+
+    let node = TestEvents::with_addr(main);
+    let other_node = TestEvents::with_addr(other);
+
+    let mut node = t1.spawn(node, connect_list.clone());
+    let other_node = t2.spawn(other_node, connect_list.clone());
 
     let message = raw_message(11, 1000);
     other_node.send_to(main, message.clone()); // should connect before send message
 
-    assert_eq!(node.wait_for_connect(), connect_message(other));
+    assert_eq!(node.wait_for_connect(), t2.connect);
     assert_eq!(node.wait_for_message(), message);
 }